model_backend/requirements.txt
fastapi==0.95.0
mangum==0.17.0
torch==2.0.0 
torchvision==0.15.1
pillow==9.5.0
python-multipart==0.0.6
scikit-learn==1.2.2
numpy==1.24.2